pyproject.toml
[project]
name = "assemblybots"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=0.27.0",
    "anthropic>=0.40.0",
    "bitsandbytes>=0.43.0",
    "click>=8.1.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-huggingface>=0.1.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-sqlite>=1.0.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "rich>=13.7.0",
    "sentencepiece>=0.1.99",
    "torch>=2.1.0",
    "tqdm>=4.66.0",
    "transformers>=4.45.0",
]
